split out footnotes
